move Operator to models
