+++ lib/src/main/java/growthbook/sdk/java/models/Operator.java
@@ -1,4 +1,4 @@
-package growthbook.sdk.java.services;
+package growthbook.sdk.java.models;
 
 import javax.annotation.Nullable;
